BXC-4062 move and rename files so they can be used for new thumbnail actions
